# Bugfix/speed up t4