--- a/timemachine/cpp/src/jitify.hpp
+++ b/timemachine/cpp/src/jitify.hpp
@@ -2465,7 +2465,7 @@ inline void detect_and_add_cuda_arch(std::vector<std::string>& options) {
   options.push_back("-arch=compute_" + ss.str());
 }
 
-inline void detect_and_add_cxx11_flag(std::vector<std::string>& options) {
+inline void detect_and_add_cxx17_flag(std::vector<std::string>& options) {
   // Reverse loop so we can erase on the fly.
   for (int i = (int)options.size() - 1; i >= 0; --i) {
     if (options[i].find("-std=c++98") != std::string::npos) {
@@ -2477,9 +2477,9 @@ inline void detect_and_add_cxx11_flag(std::vector<std::string>& options) {
       return;
     }
   }
-  // Jitify must be compiled with C++11 support, so we default to enabling it
+  // Jitify must be compiled with C++17 support, so we default to enabling it
   // for the JIT-compiled code too.
-  options.push_back("-std=c++11");
+  options.push_back("-std=c++17");
 }
 
 inline void split_compiler_and_linker_options(
@@ -2752,7 +2752,7 @@ inline void load_program(std::string const& cuda_source,
   // for arch-dependent compilation, e.g., some intrinsics are only
   // present for specific architectures.
   detail::detect_and_add_cuda_arch(compiler_options);
-  detail::detect_and_add_cxx11_flag(compiler_options);
+  detail::detect_and_add_cxx17_flag(compiler_options);
 
   // Iteratively try to compile the sources, and use the resulting errors to
   // identify missing headers.
@@ -3527,7 +3527,7 @@ Kernel_impl::Kernel_impl(Program_impl const& program, std::string name,
   _options.insert(_options.end(), _program.options().begin(),
                   _program.options().end());
   detail::detect_and_add_cuda_arch(_options);
-  detail::detect_and_add_cxx11_flag(_options);
+  detail::detect_and_add_cxx17_flag(_options);
   std::string options_string = reflection::reflect_list(_options);
   using detail::hash_combine;
   using detail::hash_larson64;
@@ -4110,7 +4110,7 @@ class KernelInstantiation {
     options.insert(options.begin(), kernel._options.begin(),
                    kernel._options.end());
     detail::detect_and_add_cuda_arch(options);
-    detail::detect_and_add_cxx11_flag(options);
+    detail::detect_and_add_cxx17_flag(options);
 
     std::string log, ptx, mangled_instantiation;
     std::vector<std::string> linker_files, linker_paths;