By using the Ampere optimization for real to int64 on the T4 GPUs, see about a 50% improvement on average time for the k_non_bonded_unified kernel

OLD
---
```
==4963== Profiling application: /home/ubuntu/.conda/envs/timemachine/bin/python /home/ubuntu/.conda/envs/timemachine/bin/pytest slow_tests/test_benchmark.py -k dhfr -s                                                                                                                    
==4963== Profiling result:                                                                                                                                                                                                                                                                 
            Type  Time(%)      Time     Calls       Avg       Min       Max  Name                                                                                                                                                                                                          
 GPU activities:   75.30%  410.75ms       297  1.3830ms  910.36us  2.4729ms  void k_nonbonded_unified<float, bool=0, bool=1, bool=0, bool=0>(int, double const *, double const *, double const *, double const *, double const *, double const *, double, double, double, int const *, unsi
gned int const *, __int64*, __int64*, __int64*, __int64*)                                                                                                                                                                                                                                  
                    4.90%  26.750ms        44  607.95us  401.89us  1.0958ms  void k_find_blocks_with_ixns<float>(int, double const *, double const *, double const *, double const *, unsigned int*, int*, unsigned int*, unsigned int*, double)                                           
                    2.76%  15.073ms       300  50.243us  32.640us  89.568us  void gen_sequenced<curandStateXORWOW, double2, normal_args_double_st, __operator_&__(double2 curand_normal_scaled2_double<curandStateXORWOW>(curandStateXORWOW*, normal_args_double_st)), rng_config<curandSta
teXORWOW, curandOrdering=101>>(curandStateXORWOW*, double2*, unsigned long, unsigned long, normal_args_double_st)                                                                                                                                                                          
                    2.67%  14.567ms        17  856.88us  1.0560us  13.879ms  [CUDA memcpy HtoD]                                                                                                                                                                                            
                    2.46%  13.426ms       303  44.311us  23.168us  73.119us  void k_nonbonded_exclusions<float>(int, double const *, double const *, double const *, double const *, double const *, double const *, double, int const *, double const *, double, double, __int64*, __int64
*, __int64*, __int64*)
```

OPTIMIZED
---------
```
==5650== Profiling application: /home/ubuntu/.conda/envs/timemachine/bin/python /home/ubuntu/.conda/envs/timemachine/bin/pytest slow_tests/test_benchmark.py -k dhfr -s                                                                                                                    
==5650== Profiling result:                                                                                                                                                                                                                                                                 
            Type  Time(%)      Time     Calls       Avg       Min       Max  Name                                                                                                                                                                                                          
 GPU activities:   65.38%  259.82ms       297  874.81us  605.12us  1.4670ms  void k_nonbonded_unified<float, bool=0, bool=1, bool=0, bool=0>(int, double const *, double const *, double const *, double const *, double const *, double const *, double, double, double, int const *, unsi
gned int const *, __int64*, __int64*, __int64*, __int64*)                                                                                                                                                                                                                                  
                    7.13%  28.319ms        44  643.61us  446.37us  1.1013ms  void k_find_blocks_with_ixns<float>(int, double const *, double const *, double const *, double const *, unsigned int*, int*, unsigned int*, unsigned int*, double)                                           
                    4.03%  16.033ms       300  53.443us  36.191us  89.599us  void gen_sequenced<curandStateXORWOW, double2, normal_args_double_st, __operator_&__(double2 curand_normal_scaled2_double<curandStateXORWOW>(curandStateXORWOW*, normal_args_double_st)), rng_config<curandSta
teXORWOW, curandOrdering=101>>(curandStateXORWOW*, double2*, unsigned long, unsigned long, normal_args_double_st)                                                                                                                                                                          
                    3.68%  14.622ms        17  860.14us  1.0240us  13.935ms  [CUDA memcpy HtoD]                                                                                                                                                                                            
                    3.41%  13.554ms       303  44.731us  22.368us  69.119us  void k_nonbonded_exclusions<float>(int, double const *, double const *, double const *, double const *, double const *, double const *, double, int const *, double const *, double, double, __int64*, __int64
*, __int64*, __int64*)
```